moved element into analysis.model, changed atomic_weight to atomic_mass
